Struct bdk_electrum::ElectrumUpdate

pub struct ElectrumUpdate<K, P> {
    pub chain_update: SparseChain<P>,
    pub last_active_indices: BTreeMap<K, u32>,
}

The result of ElectrumExt::scan.

Fields
chain_update: SparseChain<P>

The internal SparseChain update.

last_active_indices: BTreeMap<K, u32>

The last keychain script pubkey indices, which had transaction histories.

Implementations

impl<K: Ord + Clone + Debug, P: ChainPosition> ElectrumUpdate<K, P>

pub fn missing_full_txs<G>(&self, graph: G) -> Vec<&Txid>where
G: AsRef<TxGraph>,

Return a list of missing full transactions that are required to inflate_update.

pub fn into_keychain_scan<CG>(
self,
new_txs: Vec<Transaction>,
chain_graph: &CG
) -> Result<KeychainScan<K, P>, NewError<P>>where
CG: AsRef<ChainGraph<P>>,

Transform the ElectrumUpdate into a KeychainScan, which can be applied to a tracker.

This will fail if there are missing full transactions not provided via new_txs.

impl<K: Ord + Clone + Debug> ElectrumUpdate<K, TxHeight>

pub fn into_confirmation_time_update(
self,
client: &Client
) -> Result<ElectrumUpdate<K, ConfirmationTime>, Error>

Creates ElectrumUpdate<K, ConfirmationTime> from ElectrumUpdate<K, TxHeight>.