Trait bdk_electrum::ElectrumExt

pub trait ElectrumExt {
    // Required methods
    fn full_scan<K: Ord + Clone>(
        &self,
        request: FullScanRequest<K>,
        stop_gap: usize,
        batch_size: usize,
        fetch_prev_txouts: bool
    ) -> Result<ElectrumFullScanResult<K>, Error>;
    fn sync(
        &self,
        request: SyncRequest,
        batch_size: usize,
        fetch_prev_txouts: bool
    ) -> Result<ElectrumSyncResult, Error>;
}

Trait to extend [electrum_client::Client] functionality.

Required Methods

fn full_scan<K: Ord + Clone>( &self, request: FullScanRequest<K>, stop_gap: usize, batch_size: usize, fetch_prev_txouts: bool ) -> Result<ElectrumFullScanResult<K>, Error>

Full scan the keychain scripts specified with the blockchain (via an Electrum client) and returns updates for bdk_chain data structures.

request: struct with data required to perform a spk-based blockchain client full scan, see FullScanRequest

stop_gap: the full scan for each keychain stops after a gap of script pubkeys with no associated transactions

batch_size: specifies the max number of script pubkeys to request for in a single batch request

fetch_prev_txouts: specifies whether or not we want previous TxOuts for fee calculation

fn sync( &self, request: SyncRequest, batch_size: usize, fetch_prev_txouts: bool ) -> Result<ElectrumSyncResult, Error>

Sync a set of scripts with the blockchain (via an Electrum client) for the data specified and returns updates for bdk_chain data structures.

request: struct with data required to perform a spk-based blockchain client sync, see SyncRequest

batch_size: specifies the max number of script pubkeys to request for in a single batch request

fetch_prev_txouts: specifies whether or not we want previous TxOuts for fee calculation

If the scripts to sync are unknown, such as when restoring or importing a keychain that may include scripts that have been used, use full_scan with the keychain.

Object Safety

This trait is not object safe.