Trait bdk_electrum::ElectrumExt

pub trait ElectrumExt {
    fn full_scan<K: Ord + Clone>(
        &self,
        prev_tip: CheckPoint,
        keychain_spks: BTreeMap<K, impl IntoIterator<Item = (u32, ScriptBuf)>>,
        stop_gap: usize,
        batch_size: usize
    ) -> Result<(ElectrumUpdate, BTreeMap<K, u32>), Error>;
    fn sync(
        &self,
        prev_tip: CheckPoint,
        misc_spks: impl IntoIterator<Item = ScriptBuf>,
        txids: impl IntoIterator<Item = Txid>,
        outpoints: impl IntoIterator<Item = OutPoint>,
        batch_size: usize
    ) -> Result<ElectrumUpdate, Error>;
}

Trait to extend [Client] functionality.

Required Methods

fn full_scan<K: Ord + Clone>(
&self,
prev_tip: CheckPoint,
keychain_spks: BTreeMap<K, impl IntoIterator<Item = (u32, ScriptBuf)>>,
stop_gap: usize,
batch_size: usize
) -> Result<(ElectrumUpdate, BTreeMap<K, u32>), Error>

Full scan the keychain scripts specified with the blockchain (via an Electrum client) and returns updates for bdk_chain data structures.

prev_tip: the most recent blockchain tip present locally

keychain_spks: keychains that we want to scan transactions for

The full scan for each keychain stops after a gap of stop_gap script pubkeys with no associated transactions. batch_size specifies the max number of script pubkeys to request for in a single batch request.

fn sync(
&self,
prev_tip: CheckPoint,
misc_spks: impl IntoIterator<Item = ScriptBuf>,
txids: impl IntoIterator<Item = Txid>,
outpoints: impl IntoIterator<Item = OutPoint>,
batch_size: usize
) -> Result<ElectrumUpdate, Error>

Sync a set of scripts with the blockchain (via an Electrum client) for the data specified and returns updates for bdk_chain data structures.

prev_tip: the most recent blockchain tip present locally

misc_spks: an iterator of scripts we want to sync transactions for

txids: transactions for which we want updated Anchors

outpoints: transactions associated with these outpoints (residing, spending) that we want to include in the update

batch_size specifies the max number of script pubkeys to request for in a single batch request.

If the scripts to sync are unknown, such as when restoring or importing a keychain that may include scripts that have been used, use full_scan with the keychain.